Move backend api call to frontend api
